+++ src/LaryrServiceProvider.php
@@ -2,7 +2,7 @@
 
 namespace Siokas\Laryr;
 
-use Illuminate\Routing\Router;
+use Config;
 use Illuminate\Support\ServiceProvider;
 use Siokas\Laryr\Laryr;
 
@@ -21,8 +21,7 @@ class LaryrServiceProvider extends ServiceProvider
      */
     public function boot()
     {
-        // use this if your package has routes
-        $this->setupRoutes($this->app->router);
+        require __DIR__ . '/Http/routes.php';
 
         $this->publishes([
             __DIR__ . '/config/laryr.php' => config_path('laryr.php'),
@@ -33,20 +32,7 @@ class LaryrServiceProvider extends ServiceProvider
         ]);
 
     }
-    /**
-     * Define the routes for the application.
-     *
-     * @param  \Illuminate\Routing\Router  $router
-     * @return void
-     */
-    public function setupRoutes(Router $router)
-    {
-        // $router->group(['namespace' => 'Siokas\Laryr\Http\Controllers'], function ($router) {
-        //     require __DIR__ . '/Http/routes.php';
-        // });
-        require __DIR__ . '/Http/routes.php';
-        // $laryr = new Laryr();
-    }
+
     /**
      * Register any package services.
      *
@@ -54,17 +40,9 @@ class LaryrServiceProvider extends ServiceProvider
      */
     public function register()
     {
-        $this->registerLaryr();
-
         config([
             'config/laryr.php',
         ]);
     }
 
-    private function registerLaryr()
-    {
-        $this->app->bind('laryr', function ($app) {
-            return new Laryr();
-        });
-    }
 }