+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # Changelog
 
-All Notable changes to `:package_name` will be documented in this file.
+All Notable changes to `siokas/laryr` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
 ## [0.1.0] - 2016-08-01
 
 ### Added
-- Nothing
+- Yaml Routes for Laravel PHP Framework